+++ CHANGES.rst
@@ -6,6 +6,7 @@ Luftdatenpumpe changelog
 in progress
 ===========
 - Add GIS capabilities through PostGIS
+- Set default format for "stream://" targets to "json"
 
 
 2019-01-22 0.8.2