+++ CHANGES.rst
@@ -9,6 +9,7 @@ in progress
 - Set default format for "stream://" targets to "json"
 - Fix published messages getting lost when not starting
   the MQTT main loop after connecting to MQTT broker
+- Refactor station list filter
 
 
 2019-01-22 0.8.2